requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
python-dotenv>=1.0.0
chromadb>=0.4.0
faker>=24.0.0
